Switch train_rtx5090.py to native FP8 (E4M3) precision and batch size 2048

+++ train_rtx5090.py
@@ -27,7 +27,7 @@ from tensorboardX import SummaryWriter
 
 
 class AlexNet5090(nnx.Module):
-    def __init__(self, num_classes: int = 1000, dtype=jnp.bfloat16, rngs: nnx.Rngs = None):
+    def __init__(self, num_classes: int = 1000, dtype=jnp.float8_e4m3fn, rngs: nnx.Rngs = None):
         if rngs is None:
             rngs = nnx.Rngs(0)
         self.dtype = dtype
@@ -151,7 +151,7 @@ def main():
     )
 
     rngs = nnx.Rngs(0)
-    model = AlexNet5090(num_classes=1000, dtype=jnp.bfloat16, rngs=rngs)
+    model = AlexNet5090(num_classes=1000, dtype=jnp.float8_e4m3fn, rngs=rngs)
     
     # Warmup + Piecewise constant decay schedule
     warmup_steps = (len(dataset) // args.batch_size) * args.warmup_epochs
